requirements.txt
langchain 
langchain-core
langchain-openai
openai
python-dotenv
numpy 
scikit-learn